Start v2
* Remove Primer
* Use less
* add storybook
* remove old pages
