+++ .gitignore
@@ -37,3 +37,6 @@ yarn-error.log*
 # Editors
 .idea
 .vscode
+
+# Storybook
+storybook-static
