--- components/Message.module.css
@@ -1,4 +1,0 @@
-.username:hover {
-    text-decoration: underline;
-    cursor: pointer;
-}
