+++ components/index.ts
@@ -1,0 +1,5 @@
+import { Button } from "./Button";
+
+export {
+    Button,
+}
