+++ next.config.js
@@ -1,4 +1,9 @@
+import withLess from "next-with-less";
+
 /** @type {import('next').NextConfig} */
-module.exports = {
+const config = {
   reactStrictMode: true,
+  swcMinify: true,
 };
+
+module.exports = withLess(config);
