+++ package.json
@@ -8,42 +8,30 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "prettier": "prettier --write ."
+    "prettier": "prettier --write .",
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.15",
-    "@octokit/auth-token": "^2.5.0",
     "@octokit/rest": "^18.11.4",
-    "@primer/components": "^30.0.0",
     "@primer/octicons-react": "^16.0.0",
     "@supabase/supabase-js": "^1.24.0",
-    "@types/date-fns": "^2.6.0",
-    "@types/ramda": "^0.27.45",
     "country-code-emoji": "^2.3.0",
-    "date-fns": "^2.25.0",
-    "deepmerge": "^4.2.2",
-    "leven": "^4.0.0",
-    "next": "11.1.2",
-    "p-debounce": "^4.0.0",
-    "ramda": "^0.27.1",
+    "next": "^12.0.4",
+    "next-with-less": "^2.0.2",
     "react": "^17.0.2",
-    "react-div-100vh": "^0.6.0",
     "react-dom": "^17.0.2",
-    "react-hook-form": "^7.17.0",
-    "react-query": "^3.24.5",
-    "react-select": "^5.1.0",
-    "react-string-replace": "^0.4.4",
-    "react-twemoji": "^0.3.0",
-    "styled-components": "^5.3.1",
-    "uuid": "^8.3.2"
+    "react-twemoji": "^0.3.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.16.0",
+    "@storybook/addon-actions": "^6.3.12",
+    "@storybook/addon-essentials": "^6.3.12",
+    "@storybook/addon-links": "^6.3.12",
+    "@storybook/react": "^6.3.12",
     "@types/react": "17.0.26",
-    "@types/uuid": "^8.3.1",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "babel-loader": "^8.2.3",
     "eslint": "7.32.0",
     "eslint-config-next": "11.1.2",
     "eslint-config-prettier": "^8.3.0",
