+++ styles/emoji.less
@@ -1,0 +1,20 @@
+img.emoji {
+    height: 1em;
+    width: 1em;
+    margin: 0 0.05em 0 0.1em;
+    vertical-align: -0.1em;
+}
+
+img.emoji-medium {
+    height: 1.5em;
+    width: 1.5em;
+    margin: 0 0.05em 0 0.1em;
+    vertical-align: -0.1em;
+}
+
+img.emoji-large {
+    height: 2.5em;
+    width: 2.5em;
+    margin: 0 0.05em 0 0.1em;
+    vertical-align: -0.1em;
+}
