+++ styles/reset.less
@@ -1,0 +1,6 @@
+html,
+body,
+#__next {
+  background-color: #0d1117;
+  width: 100%;
+}
