+++ types/react-twemoji.d.ts
@@ -5,8 +5,8 @@ declare module 'react-twemoji' {
         tag?: string;
         noWrapper?: boolean;
         options?: {
-            callback?: Function,   // default the common replacer
-            attributes?: Function, // default returns {}
+            callback?: () => void,   // default the common replacer
+            attributes?: () => void, // default returns {}
             base?: string,         // default MaxCDN
             ext?: string,          // default ".png"
             className?: string,    // default "emoji"
